backend/requirements.txt
Pillow
tensorflow
keras
flask
flask_cors
matplotlib
torch
opencv-python